chore: seperate LinkTitle as its own component and minor fix
