+++ packages/wallets/src/components/OptionsAndMultipliersListing/OptionsAndMultipliersListing.tsx
@@ -4,7 +4,6 @@ import { useHistory } from 'react-router-dom';
 import { useActiveLinkedToTradingAccount } from '@deriv/api-v2';
 import { LabelPairedChevronRightCaptionRegularIcon } from '@deriv/quill-icons';
 import { optionsAndMultipliersContent } from '../../constants/constants';
-import { getStaticUrl, getUrlBinaryBot, getUrlSmartTrader } from '../../helpers/urls';
 import useDevice from '../../hooks/useDevice';
 import { TRoute } from '../../routes/Router';
 import { TSubscribedBalance } from '../../types';
@@ -12,48 +11,8 @@ import { WalletLink, WalletText } from '../Base';
 import { DerivAppsSection } from '../DerivAppsSection';
 import { TradingAccountCard } from '../TradingAccountCard';
 import { WalletMarketIcon } from '../WalletMarketIcon';
+import LinkTitle from './LinkTitle';
 import './OptionsAndMultipliersListing.scss';
-import { object } from 'yup';
-
-const LinkTitle: React.FC<{ platform: ComponentProps<typeof WalletMarketIcon>['icon'] }> = ({ platform }) => {
-    const handleClick = (event: React.KeyboardEvent<HTMLDivElement> | React.MouseEvent<HTMLDivElement>) => {
-        event.persist();
-        switch (platform) {
-            case 'trader':
-                window.open(getStaticUrl(`/dtrader`));
-                break;
-            case 'bot':
-                window.open(getStaticUrl(`/dbot`));
-                break;
-            case 'smarttrader':
-                window.open(getUrlSmartTrader());
-                break;
-            case 'binarybot':
-                window.open(getUrlBinaryBot());
-                break;
-            case 'derivgo':
-                window.open(getStaticUrl('/deriv-go'));
-                break;
-            default:
-                break;
-        }
-    };
-
-    return (
-        <div
-            className='wallets-options-and-multipliers-listing__content__icon'
-            onClick={handleClick}
-            // Fix sonarcloud issue
-            onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => {
-                if (event.key === 'Enter') {
-                    handleClick(event);
-                }
-            }}
-        >
-            <WalletMarketIcon icon={platform} size='lg' />
-        </div>
-    );
-};
 
 const OptionsAndMultipliersListing: React.FC<TSubscribedBalance> = ({ balance }) => {
     const { isMobile } = useDevice();
